Do not have each transformation function return a tuple
It turns out that the second element in the tuple is useless.

dbus-python encodes variants using a variant level index, rather than
a special type. Byte(32, variant_level=1) indicates that the
value had signature 'v' and is a tuple: ('y', 32).
Byte(32, variant_level=2) indicates that the value had a signature, 'v'.
The value itself is a tuple: ('v', ('y', 32)), and so forth.

The previous, more complicated code, implemented a more complicated,
incorrect indexing scheme. The idea was that the index would indicate the
entire depth of nesting, from the top of the value. This would result in
an incorrect encoding for any variant values transferred across the
D-Bus.

For example, in the previous version, a variant which was actually an
array of variants would require its elements to have a nesting index of
2. So, the signature is 'v', the value is [('v', ('y', 32))]. The
translation should be Array([Byte(32, variant_level=1)], variant_level=1)
since the array and the one element are both variants. The Byte element
should not have variant level of 2. That would indicate a value of
[('v', ('v', ('y', 32)))].

Also, remove an unnecessary test function.

Signed-off-by: mulhern <amulhern@redhat.com>