Add target for applying Monkeytype annotations
Annotate modules with what seem to be correct types and that pass the
linter and all tests.

Add a check for type annotations changes that only runs when merged, since
this package is rather dormant and so nightly is annoyingly frequent.
Restrict its operation to modules where it does not currently introduce
annotations that cause some kind of test failure, or make syntactic changes
in the annotations that result in a diff.

This reduces the number of modules automatically checked to exactly one,
but other projects are likely to do a bit better than that.

Signed-off-by: mulhern <amulhern@redhat.com>